Gigantic refactor session, mostly to the DataManager class to centalize data-centric code.
All data is loaded on the index page. (May be better to load it on login instead.) To reduce the amounts of calls to database.

Everything is then updated locally and data is only sent to the database to again reduce database traffic.

Refactored tables to make the related pages more readable.

Will likely need to refactor pages that aren't listed in this commit too. But this should make the whole repo more readable.